pyproject.toml
[tool.poetry]
name = "standardizer"
version = "0.1.0"
description = ""
authors = ["Pierre Merienne"]
readme = "README.md"
packages = [{ include = "standardizer" }]

[tool.poetry.dependencies]
python = "^3.10"
gradio = "^5.4.0"
haystack-ai = "^2.6.1"
loguru = "^0.7.2"
chroma-haystack = "^0.22.1"
langfuse-haystack = "^0.5.0"
requests = "^2.32.3"
notion-haystack = "^1.0.0"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
standardizer = "standardizer.main:launch_ui"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 160
ignore = [
    "E402", # Module level import not at top of file
]
